Project 2 - Student Survey
This project was my first taste of analyzing data and creating a visualization based on what I discovered.  The data was an Excel spreadsheet with a sampling of Udacity Student Surveys.  I was required to come up with four questions and provide visualizations to accompany the analysis.  There is a correction I noted on the 2nd slide of my presentation of something to consider for future projects.  I left it there as a reminder and to remember for future projects.